+++ src/main/java/co/programmers/voucher_management/common/Response.java
@@ -1,4 +1,4 @@
-package co.programmers.voucher_management;
+package co.programmers.voucher_management.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
